requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-whois==0.8.0
requests==2.31.0
tqdm==4.66.1
whois21==1.4.6
pypinyin==0.49.0 